requirements/base.txt
Django==1.6.2
Unipath==1.0
djangorestframework==2.3.13
django-cors-headers==0.12
django-mptt==0.6.0
